// types/carriers.ts
export interface Carrier {
  ADDRL1: string
  ADDRL2: string
  CARDSC: string
  CARID: string
  CHGDAT: string
  CHGTIM: string
  CHGUSR: string
  CNTYCOD: string
  COMPID: string
  CONNME: string
  CRTDAT: string
  CRTTIM: string
  CRTUSR: string
  City: string
  Fax: string
  Phone: string
  RowNum: number
  STAID: string
  WebSite: string
  ZIPCODE: string
  eMail: string
} 